Move Handle wrapper to own pub mod
